requirements.txt
pytest
cocotb
cocotb-test
GitPython
pytest-xdist[mp]
setproctitle[mp]